pyproject.toml
[project]
name = "django-rest-knox-redis"
version = "0.1.0"
description = "Redis caching layer for django-rest-knox - dramatically reduce database load on token authentication"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your@email.com"}
]
keywords = [
    "django",
    "rest",
    "knox",
    "redis",
    "authentication",
    "token",
    "cache",
    "performance",
    "scaling",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Session",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
    "django-rest-knox>=4.2.0",
    "django-redis>=5.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.5",
    "pytest-cov>=4.0",
    "pytest-mock>=3.12",
    "fakeredis>=2.20",
    "ruff>=0.1.0",
    "mypy>=1.8",
    "django-stubs>=4.2",
    "djangorestframework-stubs>=3.14",
]

[project.urls]
Homepage = "https://github.com/yourusername/django-rest-knox-redis"
Documentation = "https://github.com/yourusername/django-rest-knox-redis#readme"
Repository = "https://github.com/yourusername/django-rest-knox-redis"
Issues = "https://github.com/yourusername/django-rest-knox-redis/issues"
Changelog = "https://github.com/yourusername/django-rest-knox-redis/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["knox_redis"]

[tool.hatch.build.targets.sdist]
include = [
    "/knox_redis",
    "/tests",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.5",
    "pytest-cov>=4.0",
    "pytest-mock>=3.12",
    "fakeredis>=2.20",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["knox_redis"]

[tool.mypy]
python_version = "3.10"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.django-stubs]
django_settings_module = "tests.settings"

[tool.coverage.run]
source = ["knox_redis"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
